error handling for missing (or extra) trial markers
handle marker alignment in NIRS sessions where we accidentally hit stop recording on one of the systems before quitting the task (so extra trial markers end up in either the NEV or NIRS file). this now generates a plot showing marker alignment, asks the user to confirm the number of missing markers, and then adjusts the nirs data or dat file accordingly.